Trivial: rename cxxtimer.hpp to cxxtimer.h
