+++ src/Makefile.am
@@ -163,7 +163,7 @@ BITCOIN_CORE_H = \
   clientversion.h \
   coincontrol.h \
   coins.h \
-  cxxtimer.hpp \
+  cxxtimer.h \
   compat.h \
   compat/byteswap.h \
   compat/cpuid.h \
